Update dependencies, remove code duplication
